Add jsgfkitxx subdirectory for include dirs and rename files to clarify word boundaries
