+++ examples/Matching/Matching.cpp
@@ -1,4 +1,4 @@
-#include "grammar.h"
+#include "Grammar.hpp"
 #include <iostream>
 #include <string>
