+++ include/jsgfkitxx/MatchInfo.hpp
@@ -1,6 +1,6 @@
 #ifndef MATCHINFO_H
 #define MATCHINFO_H
-#include "expansion.h"
+#include "Expansion.hpp"
 #include <string>
 
 class MatchInfo
