+++ include/jsgfkitxx/PlusOperator.hpp
@@ -1,20 +1,20 @@
-#ifndef OPTIONALGROUPING_H
-#define OPTIONALGROUPING_H
+#ifndef PLUSOPERATOR_H
+#define PLUSOPERATOR_H
 
-#include <expansion.h>
+#include "jsgfkitxx/Expansion.hpp"
 
 
-class OptionalGrouping : public Expansion
+class PlusOperator : public Expansion
 {
     private:
         std::shared_ptr<Expansion> childExpansion;
 
     public:
         /** Default constructor */
-        OptionalGrouping();
-        OptionalGrouping(std::shared_ptr<Expansion> e);
+        PlusOperator();
+        PlusOperator(std::shared_ptr<Expansion> e);
         /** Default destructor */
-        ~OptionalGrouping();
+        ~PlusOperator();
 
         Expansion * clone();
         void replaceChild(std::shared_ptr<Expansion> newChild, const unsigned long index = 0);
@@ -26,5 +26,4 @@ class OptionalGrouping : public Expansion
 
         std::string getText() const;
 };
-
-#endif // OPTIONALGROUPING_H
+#endif // PLUSOPERATOR_H
