+++ src/matchinfo.cpp
@@ -1,4 +1,4 @@
-#include "matchinfo.h"
+#include "jsgfkitxx/MatchInfo.hpp"
 
 MatchInfo::MatchInfo(std::shared_ptr<Expansion> e, std::string matchingStringPart)
 {
