+++ src/unparsedsection.cpp
@@ -1,5 +1,6 @@
-#include "unparsedsection.h"
+#include "jsgfkitxx/UnparsedSection.hpp"
 #include <iostream>
+
 UnparsedSection::UnparsedSection()
 {
     //ctor
